% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_footprint.R
\name{land_footprint}
\alias{land_footprint}
\title{Land Footprint (LF) of a Wind Turbine}
\usage{
land_footprint(rd, pwr, nt = 25, cf = 0.2, y = 20)
}
\arguments{
\item{rd}{rotor diameter (m)}

\item{pwr}{rated capacity of the turbine (MW)}

\item{nt}{total number of turbines default = 25}

\item{cf}{capacity factory of wind turbine default = 0.2}

\item{y}{Lifetime of plant (year) default = 20}
}
\value{
land footprint (m2/GWh)
}
\description{
This function determines the Land Footprint of a wind turbine based on turbine configuration, rotor
diameter and rated capacity of the turbine
}
